subject wise decomposition

+++ code/START_A2_smooth_decomposition.py
@@ -25,6 +25,11 @@ def START_A2_smooth_decomposition(dataset_name):
     output = dict()
     PKL_output = os.path.join(resultsPath, 'output.pkl')
 
+    output_subject_wise = dict()
+    PKL_output_subject_wise = os.path.join(resultsPath, 'output_subject_wise.pkl')
+
+    redo_smoothing = False
+
     # Get the atlas
     atlas_str = 'fs32k'
     atlas, ainf = am.get_atlas(atlas_str)
@@ -61,58 +66,59 @@ def START_A2_smooth_decomposition(dataset_name):
         X_individuals = X_individuals_ext
 
     ## smoothing
-    # save orig, no smoothing
-    PKL_smoothed = os.path.join(resultsPath, f'smoothed_orig_mm.pkl')
-    with open(PKL_smoothed, 'wb') as pk:
-        pickle.dump(X_individuals, pk)
-
-    n_subjects, n_conds, n_vertices = X_individuals.shape
-
-    smoothing_kernels = [10, 8, 6, 4, 2]      # mm, fwhm
-
-    X_individuals_smoothed = np.zeros(X_individuals.shape)
-
-    temp_smoothing_path = os.path.join(resultsPath, 'temp')
-    if not os.path.exists(temp_smoothing_path):
-        os.makedirs(temp_smoothing_path)
-
-    for smoothing_kernel in smoothing_kernels:
-        for subjI in np.arange(n_subjects):
-            for condI in np.arange(n_conds):
-                X_subj_cond = X_individuals[subjI, condI, :].reshape(1, -1)
-                cifti_subj_cond = atlas.data_to_cifti(X_subj_cond)
-                cifti_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d.dscalar.nii' % (subjI, condI))
-                nib.save(cifti_subj_cond, cifti_file_name)
-
-                cifti_smoothed_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d_smoothed_%dmm.dscalar.nii' % (subjI, condI, smoothing_kernel))
-                wb_cmd = f'wb_command -cifti-smoothing {cifti_file_name} {smoothing_kernel} {smoothing_kernel} COLUMN {cifti_smoothed_file_name} -left-surface {flat_surf_L} -right-surface {flat_surf_R} -fwhm'
-                subprocess.run(wb_cmd, shell=True)
-
-                cifti_subj_cond_smoothed = nib.load(cifti_smoothed_file_name)
-                X_individuals_smoothed[subjI, condI, :] = atlas.cifti_to_data(cifti_subj_cond_smoothed)
-
-        X_individuals_smoothed[np.where(np.isnan(X_individuals_smoothed))] = 0
-
-        PKL_smoothed = os.path.join(resultsPath, f'smoothed_{smoothing_kernel}_mm.pkl')
+    if redo_smoothing:
+        # save orig, no smoothing
+        PKL_smoothed = os.path.join(resultsPath, f'smoothed_orig_mm.pkl')
         with open(PKL_smoothed, 'wb') as pk:
-            pickle.dump(X_individuals_smoothed, pk)
+            pickle.dump(X_individuals, pk)
 
-        for subjI in np.arange(n_subjects):
-            for condI in np.arange(n_conds):
-                X_individuals[subjI, condI, :] = X_individuals[subjI, condI, :] - X_individuals_smoothed[subjI, condI, :]
+        n_subjects, n_conds, n_vertices = X_individuals.shape
 
-                cifti_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d.dscalar.nii' % (subjI, condI))
-                if os.path.exists(cifti_file_name):
-                    os.remove(cifti_file_name)
+        smoothing_kernels = [10, 8, 6, 4, 2]      # mm, fwhm
 
-                cifti_smoothed_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d_smoothed_%dmm.dscalar.nii' % (subjI, condI, smoothing_kernel))
-                if os.path.exists(cifti_smoothed_file_name):
-                    os.remove(cifti_smoothed_file_name)
+        X_individuals_smoothed = np.zeros(X_individuals.shape)
 
-    # save residuals (finer than 2mm)
-    PKL_smoothed = os.path.join(resultsPath, f'smoothed_residuals_mm.pkl')
-    with open(PKL_smoothed, 'wb') as pk:
-        pickle.dump(X_individuals, pk)
+        temp_smoothing_path = os.path.join(resultsPath, 'temp')
+        if not os.path.exists(temp_smoothing_path):
+            os.makedirs(temp_smoothing_path)
+
+        for smoothing_kernel in smoothing_kernels:
+            for subjI in np.arange(n_subjects):
+                for condI in np.arange(n_conds):
+                    X_subj_cond = X_individuals[subjI, condI, :].reshape(1, -1)
+                    cifti_subj_cond = atlas.data_to_cifti(X_subj_cond)
+                    cifti_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d.dscalar.nii' % (subjI, condI))
+                    nib.save(cifti_subj_cond, cifti_file_name)
+
+                    cifti_smoothed_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d_smoothed_%dmm.dscalar.nii' % (subjI, condI, smoothing_kernel))
+                    wb_cmd = f'wb_command -cifti-smoothing {cifti_file_name} {smoothing_kernel} {smoothing_kernel} COLUMN {cifti_smoothed_file_name} -left-surface {flat_surf_L} -right-surface {flat_surf_R} -fwhm'
+                    subprocess.run(wb_cmd, shell=True)
+
+                    cifti_subj_cond_smoothed = nib.load(cifti_smoothed_file_name)
+                    X_individuals_smoothed[subjI, condI, :] = atlas.cifti_to_data(cifti_subj_cond_smoothed)
+
+            X_individuals_smoothed[np.where(np.isnan(X_individuals_smoothed))] = 0
+
+            PKL_smoothed = os.path.join(resultsPath, f'smoothed_{smoothing_kernel}_mm.pkl')
+            with open(PKL_smoothed, 'wb') as pk:
+                pickle.dump(X_individuals_smoothed, pk)
+
+            for subjI in np.arange(n_subjects):
+                for condI in np.arange(n_conds):
+                    X_individuals[subjI, condI, :] = X_individuals[subjI, condI, :] - X_individuals_smoothed[subjI, condI, :]
+
+                    cifti_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d.dscalar.nii' % (subjI, condI))
+                    if os.path.exists(cifti_file_name):
+                        os.remove(cifti_file_name)
+
+                    cifti_smoothed_file_name = os.path.join(temp_smoothing_path, 'cifti_sub-%02d_cond_%02d_smoothed_%dmm.dscalar.nii' % (subjI, condI, smoothing_kernel))
+                    if os.path.exists(cifti_smoothed_file_name):
+                        os.remove(cifti_smoothed_file_name)
+
+        # save residuals (finer than 2mm)
+        PKL_smoothed = os.path.join(resultsPath, f'smoothed_residuals_mm.pkl')
+        with open(PKL_smoothed, 'wb') as pk:
+            pickle.dump(X_individuals, pk)
 
 
     ## decomposition
@@ -127,6 +133,7 @@ def START_A2_smooth_decomposition(dataset_name):
 
     for smoothing_kernel in smoothing_kernels:
         output[f'{smoothing_kernel}_mm'] = {}
+        output_subject_wise[f'{smoothing_kernel}_mm'] = {}
 
         PKL_smoothed = os.path.join(resultsPath, f'smoothed_{smoothing_kernel}_mm.pkl')
         with open(PKL_smoothed, 'rb') as pf:
@@ -138,8 +145,14 @@ def START_A2_smooth_decomposition(dataset_name):
             variances = decompose_subj_group(data_region, cond_vec, part_vec, separate='none')
             output[f'{smoothing_kernel}_mm'][roi] = variances
 
+            variances = decompose_subj_group(data_region, cond_vec, part_vec, separate='subject_wise')
+            output_subject_wise[f'{smoothing_kernel}_mm'][roi] = variances
+
     with open(PKL_output, 'wb') as pk:
         pickle.dump(output, pk)
+
+    with open(PKL_output_subject_wise, 'wb') as pk:
+        pickle.dump(output_subject_wise, pk)
 
 
 if __name__=='__main__':
@@ -148,5 +161,8 @@ if __name__=='__main__':
     # for dataset in datasets:
     #     START_A2_smooth_decomposition(dataset_name=dataset)
 
+
     if len(sys.argv) > 1:
         START_A2_smooth_decomposition(sys.argv[1])
+    else:
+        START_A2_smooth_decomposition('MDTB')
